requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CrewAI
crewai==1.9.2
crewai-tools==1.9.2

# LiteLLM (LLM wrapper - no vendor lock-in)
litellm==1.56.3

# Spotify
spotipy==2.23.0

# Genius API
lyricsgenius==3.0.1

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# PDF Generation
reportlab==4.0.9
pypdf==4.0.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-multipart==0.0.6